HTML5 emsdk troubleshooting notes
add details on possible issues when emsdk might go wrong (which is run
during HTML5Setup.sh).

this includes some suggestions on how to fix this.

a better solution would be to edit HTML5Setup.sh to double check
that all "tools" folders are installed.  and to rerun "emsdk install"
(instead of just checking if emsdk was installed).
- or remove the the DO_INSTALL check (and "always" run "emsdk install")

resolves #2